fix Patient and Employee and create Token middleware methode for this

+++ controllers/employeesController.js
@@ -7,6 +7,16 @@ const { jwtSecret, jwtExpiration } = require('../middlewares/tockenService');
 
 async function getAll(req, res) {
     try {
+
+        const token = req.headers['authorization'].split(' ')[1];
+
+        const employeeTokenId = await EmployeeToken.findOne({ where: { token } });
+
+
+        if (!employeeTokenId) {
+            return res.status(404).json({ error: 'employeeTokenId not found!' });
+        }
+
         const employees = await Employee.findAll();
         if (employees.length === 0) {
             return res.status(404).json({ error: 'Keine Employee gefunden!!' });;
@@ -42,6 +52,14 @@ async function deleteEmployee(req, res) {
     try {
         const { id } = req.query;
 
+        const token = req.headers['authorization'].split(' ')[1];
+
+        const employeeTokenId = await EmployeeToken.findOne({ where: { token } });
+
+        if (!employeeTokenId) {
+            return res.status(404).json({ error: 'employeeTokenId not found!' });
+        }
+
         const deletedEmployee = await Employee.destroy({ where: { id } });
 
         if (!deletedEmployee) {
@@ -60,6 +78,17 @@ async function updateEmployee(req, res) {
     try {
         const { id } = req.query; // Extrahieren der ID aus den Abfrageparametern
         const updates = req.body;
+
+        const token = req.headers['authorization'].split(' ')[1];
+
+        const employeeTokenId = await EmployeeToken.findOne({ where: { token } });
+
+
+        if (!employeeTokenId) {
+            return res.status(404).json({ error: 'employeeTokenId not found!' });
+        }
+
+
         if (updates.password) {
             updates.password = await bcrypt.hash(updates.password, 10);
         }
@@ -107,11 +136,36 @@ async function login(req, res) {
     }
 }
 
+async function logout(req, res) {
+    try {
+        const token = req.headers['authorization'].split(' ')[1];
+
+        // Find the patient ID using the token
+        const employeeToken = await EmployeeToken.findOne({ where: { token } });
+        if (!employeeToken) {
+            return res.status(404).json({ error: 'Employee not found!' });
+        }
+        const employee = await Employee.findOne({ where: { id: employeeToken.employeeId } });
+        if (!employee) {
+            return res.status(404).json({ error: 'Employee not found!' });
+        }
+        await EmployeeToken.destroy({ where: { token } });
+        // Update the patient's activity status to false
+        await employee.update({ active: false });
+
+        return res.status(200).json({ message: 'Logout successful!' });
+    } catch (error) {
+        console.error('Error logging out:', error);
+        return res.status(500).json({ error: 'An error occurred while logging out!' });
+    }
+}
+
 
 module.exports = {
     getAll,
     signup,
     deleteEmployee,
     updateEmployee,
-    login
+    login,
+    logout
 }
